Move wasm executor to a crate under fuel-indexer
