+++ fuel-core/Cargo.toml
@@ -21,15 +21,14 @@ axum = { version = "0.2" }
 bincode = "1.3"
 chrono = { version = "0.4", features = ["serde"] }
 derive_more = { version = "0.99" }
-diesel = { version = "1.4", features = ["postgres", "serde_json"] }
 dirs = "3.0"
 env_logger = "0.9"
 fuel-asm = { git = "ssh://git@github.com/FuelLabs/fuel-asm.git", features = ["serde-types"] }
-fuel-indexer-schema = { path = "../fuel-indexer/schema", features = ["db-models"] }
 fuel-storage = { git = "ssh://git@github.com/FuelLabs/fuel-storage.git" }
 fuel-tx = { git = "ssh://git@github.com/FuelLabs/fuel-tx.git", features = ["serde-types"] }
 fuel-types = { git = "ssh://git@github.com/FuelLabs/fuel-types.git", features = ["serde-types"] }
 fuel-vm = { git = "ssh://git@github.com/FuelLabs/fuel-vm.git", features = ["serde-types"] }
+fuel-wasm-executor = { path = "../fuel-indexer/indexer" }
 futures = "0.3"
 graphql-parser = "0.3.0"
 hex = "0.4"
@@ -38,11 +37,8 @@ itertools = "0.10"
 lazy_static = "1.4"
 rand = "0.8"
 rocksdb = { version = "0.17", features = ["snappy", "multi-threaded-cf"], optional = true }
-r2d2 = "0.8"
-r2d2-diesel = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version= "1.0", features = ["raw_value"] }
-serde-scale = "0.2"
 structopt = "0.3"
 strum = "0.21"
 strum_macros = "0.21"
@@ -53,15 +49,9 @@ tower-http = { version = "0.1", features = ["set-header"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.2", features = ["env-filter"] }
 uuid = { version = "0.8", features = ["v4"] }
-wasmer = "2.0.0"
-wasmer-engine-universal = "2.0.0"
-wasmer-compiler-llvm = "2.0.0"
 
 [dev-dependencies]
-fuel-indexer = { path = "../fuel-indexer/lib" }
 fuel-vm = { git = "ssh://git@github.com/FuelLabs/fuel-vm.git", features = ["serde-types", "random"] }
-serde_yaml = "0.8.19"
 
 [features]
 default = ["rocksdb"]
-postgres = []
