+++ fuel-indexer/indexer/src/executor.rs
@@ -1,6 +1,6 @@
-use crate::wasm_executor::database::Database;
-use crate::wasm_executor::ffi;
-use crate::wasm_executor::{IndexerError, IndexerResult, Manifest};
+use crate::database::Database;
+use crate::ffi;
+use crate::{IndexerError, IndexerResult, Manifest};
 use std::collections::HashMap;
 use std::path::Path;
 use std::sync::{Arc, Mutex};
