+++ fuel-indexer/indexer/src/lib.rs
