+++ fuel-indexer/indexer/src/manifest.rs
