+++ fuel-indexer/indexer/src/test_data/bad_manifest.yaml
