+++ fuel-indexer/indexer/src/test_data/good_event.bin
