+++ fuel-indexer/indexer/src/test_data/manifest.yaml
