+++ fuel-indexer/indexer/src/test_data/schema.graphql
