+++ fuel-indexer/indexer/src/test_data/simple_wasm.wasm
