+++ fuel-indexer/indexer/tests/test_data/manifest.yaml
