+++ fuel-indexer/indexer/tests/test_data/schema.graphql
