+++ fuel-indexer/indexer/tests/test_data/simple_wasm.wasm
